automation-saas/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg[binary]==3.2.3
python-dotenv==1.0.1
apscheduler==3.10.4
tweepy==4.15.0
httpx==0.27.0
tenacity==8.4.1
twikit>=2.3.3
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1
